requirements.txt
fastapi
uvicorn[standard]
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlite3
opentelemetry-exporter-jaeger
prometheus-client
